// src/foundation/memory/AccessorBase.d.ts
import { ComponentTypeEnum } from "../definitions/ComponentType";
import { CompositionTypeEnum } from "../definitions/CompositionType";
import RnObject from "../core/Object";
import BufferView from "./BufferView";
import Vector2 from "../math/Vector2";
import Vector3 from "../math/Vector3";
import Vector4 from "../math/Vector4";
import Matrix33 from "../math/Matrix33";
import MutableMatrix44 from "../math/MutableMatrix44";
export default class AccessorBase extends RnObject {
    protected __bufferView: BufferView;
    protected __byteOffsetInBuffer: number;
    protected __compositionType: CompositionTypeEnum;
    protected __componentType: ComponentTypeEnum;
    protected __count: Count;
    protected __raw: ArrayBuffer;
    protected __dataView?: DataView;
    protected __typedArray?: TypedArray;
    protected __takenCount: Count;
    protected __byteStride: Byte;
    protected __typedArrayClass?: TypedArrayConstructor;
    protected __dataViewGetter: any;
    protected __dataViewSetter: any;
    constructor({ bufferView, byteOffset, compositionType, componentType, byteStride, count, raw }: {
        bufferView: BufferView;
        byteOffset: Byte;
        compositionType: CompositionTypeEnum;
        componentType: ComponentTypeEnum;
        byteStride: Byte;
        count: Count;
        raw: Uint8Array;
    });
    prepare(): void;
    getTypedArrayClass(componentType: ComponentTypeEnum): TypedArrayConstructor | undefined;
    getDataViewGetter(componentType: ComponentTypeEnum): string | undefined;
    getDataViewSetter(componentType: ComponentTypeEnum): string | undefined;
    takeOne(): TypedArray;
    readonly numberOfComponents: number;
    readonly componentSizeInBytes: number;
    readonly elementSizeInBytes: number;
    readonly elementCount: Count;
    readonly byteLength: Byte;
    readonly componentType: ComponentTypeEnum;
    readonly compositionType: CompositionTypeEnum;
    getTypedArray(): TypedArray;
    readonly isAoS: boolean;
    readonly isSoA: boolean;
    readonly byteStride: number;
    getScalar(index: Index, endian?: boolean): number;
    getScalarAt(index: Index, compositionOffset: Index, endian?: boolean): number;
    getVec2AsArray(index: Index, endian?: boolean): Array<number>;
    getVec3AsArray(index: Index, endian?: boolean): Array<number>;
    getVec4AsArray(index: Index, endian?: boolean): Array<number>;
    getMat3AsArray(index: Index, endian?: boolean): Array<number>;
    getMat4AsArray(index: Index, endian?: boolean): Array<number>;
    getVec2(index: Index, endian?: boolean): Vector2;
    getVec3(index: Index, endian?: boolean): Vector3;
    getVec4(index: Index, endian?: boolean): Vector4;
    getMat3(index: Index, endian?: boolean): Matrix33;
    getMat4(index: Index, endian?: boolean): MutableMatrix44;
    setScalar(index: Index, value: number, endian?: boolean): void;
    setVec2(index: Index, x: number, y: number, endian?: boolean): void;
    setVec3(index: Index, x: number, y: number, z: number, endian?: boolean): void;
    setVec4(index: Index, x: number, y: number, z: number, w: number, endian?: boolean): void;
    copyFromTypedArray(typedArray: TypedArray): void;
    setScalarAt(index: Index, conpositionOffset: Index, value: number, endian?: boolean): void;
    readonly arrayBufferOfBufferView: ArrayBuffer;
    readonly dataViewOfBufferView: DataView;
    readonly byteOffsetInBufferView: Byte;
    readonly byteOffsetInBuffer: Byte;
    readonly bufferView: BufferView;
}